requirements.txt
flask==2.3.1
pandas==2.2.0
plotly==5.16.1
supabase==2.0.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
requests>=2.27.0
httpx>=0.24.0
postgrest==0.10.8
websockets>=11.0 